15-01-2021 Completed tour query functionality client side
