+++ jaden-website/src/components/TourQuery.js
@@ -1,0 +1,32 @@
+import React from 'react'
+import './TourQuery.css'
+import { Button } from './Button'
+
+function TourQuery(props) {
+
+    function buyTickets() {
+        alert('Tickets purchased')
+    }
+
+    if (props.networkStatus === 4) return <h1>Refetching</h1>
+    if (props.loading) return <h1>Loading...</h1>;
+    if (props.error) return <h1>Error! ${props.error}</h1>;
+
+    return (
+        props.data.searchTour.map((tour, index) => (
+            <div className="tour-item" key={index}>
+                <div className="tour-info">
+                    <strong className="tour-date">{tour.date}</strong>
+                    <span className="tour-city">{tour.city}</span>
+                    <a href={tour.link} target="_blank" rel="noreferrer" className="map-location">
+                        <i class="fas fa-map-marked-alt"></i>
+                    </a>
+                    <span className="tour-arena">{tour.arena}</span>  
+                </div>
+                <Button buttonStyle="btn--buy"      buttonSize="btn--medium" onClick={buyTickets}>BUY TICKETS</Button>
+            </div> 
+        ))
+    )
+}
+
+export default TourQuery
